Encryptors/python/requirements.txt
pycryptodome==3.14.1
ConfigParser